Refactored controllers and dtos to corressponding domens
