+++ Projekatsiit2023/Projekatsiit2023/src/main/java/rs/ac/uns/ftn/asd/Projekatsiit2023/controller/PublicMapController.java
@@ -4,7 +4,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import rs.ac.uns.ftn.asd.Projekatsiit2023.dto.ActiveVehicleDTO;
+
+import rs.ac.uns.ftn.asd.Projekatsiit2023.dto.driver.ActiveVehicleDTO;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -13,7 +14,7 @@ import java.util.List;
 @RequestMapping("/api/public-map")
 public class PublicMapController {
 
-    @GetMapping({"/active"})
+    @GetMapping({ "/active" })
     public ResponseEntity<List<ActiveVehicleDTO>> getActiveVehicles() {
         List<ActiveVehicleDTO> response = new ArrayList<>();
         response.add(new ActiveVehicleDTO());
