+++ Projekatsiit2023/Projekatsiit2023/src/main/java/rs/ac/uns/ftn/asd/Projekatsiit2023/controller/ride/RideHistoryController.java
@@ -1,8 +1,9 @@
-package rs.ac.uns.ftn.asd.Projekatsiit2023.controller;
+package rs.ac.uns.ftn.asd.Projekatsiit2023.controller.ride;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import rs.ac.uns.ftn.asd.Projekatsiit2023.dto.DriverRideHistoryDTO;
+
+import rs.ac.uns.ftn.asd.Projekatsiit2023.dto.driver.DriverRideHistoryDTO;
 
 import java.time.LocalDateTime;
 import java.util.*;
@@ -14,8 +15,8 @@ public class RideHistoryController {
     @GetMapping("/{driverId}/rides")
     public ResponseEntity<List<DriverRideHistoryDTO>> getDriverRideHistory(
             @PathVariable Long driverId,
-            @RequestParam(required = false)LocalDateTime from,
-            @RequestParam(required = false)LocalDateTime to) {
+            @RequestParam(required = false) LocalDateTime from,
+            @RequestParam(required = false) LocalDateTime to) {
         List<DriverRideHistoryDTO> response = new ArrayList<>();
         response.add(new DriverRideHistoryDTO());
         return ResponseEntity.ok(response);
