+++ Projekatsiit2023/Projekatsiit2023/src/main/java/rs/ac/uns/ftn/asd/Projekatsiit2023/dto/auth/LoginRequestDTO.java
@@ -1,10 +1,11 @@
-package rs.ac.uns.ftn.asd.Projekatsiit2023.dto;
+package rs.ac.uns.ftn.asd.Projekatsiit2023.dto.auth;
 
 public class LoginRequestDTO {
     private String email;
     private String password;
 
-    public LoginRequestDTO() {}
+    public LoginRequestDTO() {
+    }
 
     public String getEmail() {
         return email;
