+++ Projekatsiit2023/Projekatsiit2023/src/main/java/rs/ac/uns/ftn/asd/Projekatsiit2023/dto/auth/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package rs.ac.uns.ftn.asd.Projekatsiit2023.dto;
+package rs.ac.uns.ftn.asd.Projekatsiit2023.dto.auth;
 
 public class LoginResponseDTO {
     private String token;
@@ -6,7 +6,8 @@ public class LoginResponseDTO {
     private String email;
     private String role;
 
-    public LoginResponseDTO() {}
+    public LoginResponseDTO() {
+    }
 
     public String getToken() {
         return token;
