+++ Projekatsiit2023/Projekatsiit2023/src/main/java/rs/ac/uns/ftn/asd/Projekatsiit2023/dto/ride/RideRatingResponseDTO.java
@@ -1,4 +1,4 @@
-package rs.ac.uns.ftn.asd.Projekatsiit2023.dto;
+package rs.ac.uns.ftn.asd.Projekatsiit2023.dto.ride;
 
 public class RideRatingResponseDTO {
     private Long rideId;
@@ -7,7 +7,8 @@ public class RideRatingResponseDTO {
     private Long vehicleId;
     private int vehicleRating;
 
-    public RideRatingResponseDTO() {}
+    public RideRatingResponseDTO() {
+    }
 
     public Long getRideId() {
         return rideId;
