+++ Projekatsiit2023/Projekatsiit2023/src/main/java/rs/ac/uns/ftn/asd/Projekatsiit2023/dto/ride/RideTrackingResponseDTO.java
@@ -1,4 +1,4 @@
-package rs.ac.uns.ftn.asd.Projekatsiit2023.dto;
+package rs.ac.uns.ftn.asd.Projekatsiit2023.dto.ride;
 
 public class RideTrackingResponseDTO {
     private Long rideId;
@@ -6,29 +6,37 @@ public class RideTrackingResponseDTO {
     private String nextStop;
     private int timeLeft;
 
-    public RideTrackingResponseDTO() {}
+    public RideTrackingResponseDTO() {
+    }
 
     public Long getRideId() {
         return rideId;
     }
+
     public void setRideId(Long rideId) {
         this.rideId = rideId;
     }
+
     public String getCurrentLocation() {
         return currentLocation;
     }
+
     public void setCurrentLocation(String currentLocation) {
         this.currentLocation = currentLocation;
     }
+
     public String getNextStop() {
         return nextStop;
     }
+
     public void setNextStop(String nextStop) {
         this.nextStop = nextStop;
     }
+
     public int getTimeLeft() {
         return timeLeft;
     }
+
     public void setTimeLeft(int timeLeft) {
         this.timeLeft = timeLeft;
     }
